adding the router file and taking it out of init
