+++ app/__init__.py
@@ -6,12 +6,6 @@ from app.db.connection import initialize_db
 app = FastAPI()
 
 
-# Initialize routers
-def routers():
-    app.include_router(commits.router, prefix="/commits", tags=["Commits"])
-    app.include_router(files.router, prefix="/files", tags=["Files"])
-    app.include_router(repositories.router, prefix="/repositories", tags=["Repositories"])
-    app.include_router(branches.router, prefix="/branches", tags=["Branches"])
 
 
 # Initialize services (placeholder, as services are directly imported in routers)
